Migrate HouseBrowser into dedicated component
