+++ src/components/styles/App.module.scss
@@ -7,44 +7,5 @@
 
   .row {
     flex-grow: 1;
-
-    .nav-column {
-      display: flex;
-    }
-
-    .main-column {
-      display: flex;
-      flex-direction: column;
-      align-items: center;
-      justify-content: center;
-      padding-top: 100px;
-
-      h1,
-      h5 {
-        font-family: "Lato", sans-serif;
-      }
-
-      h1 {
-        margin-bottom: 0px;
-      }
-
-      h5 {
-        color: rgba(150, 150, 150, 0.8);
-        font-size: 17px;
-        margin-bottom: 40px;
-      }
-
-      .carousel {
-        width: 95%;
-
-        .carousel-indicators {
-          margin: 0;
-        }
-        .image {
-          max-width: 100%;
-          max-height: 100%;
-        }
-      }
-    }
   }
 }
