+++ src/components/styles/HouseBrowser.module.scss
@@ -1,0 +1,36 @@
+@import url("https://fonts.googleapis.com/css2?family=Lato:wght@900&display=swap");
+
+.main-column {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  padding-top: 100px;
+
+  h1,
+  h5 {
+    font-family: "Lato", sans-serif;
+  }
+
+  h1 {
+    margin-bottom: 0px;
+  }
+
+  h5 {
+    color: rgba(150, 150, 150, 0.8);
+    font-size: 17px;
+    margin-bottom: 40px;
+  }
+
+  .carousel {
+    width: 95%;
+
+    .carousel-indicators {
+      margin: 0;
+    }
+    .image {
+      max-width: 100%;
+      max-height: 100%;
+    }
+  }
+}
