+++ src/components/styles/NavBar.module.scss
@@ -1,50 +1,53 @@
 @import url("https://fonts.googleapis.com/css2?family=Lato:wght@900&display=swap");
 
-.nav-bar {
-  flex-grow: 1;
+.nav-column {
   display: flex;
-  flex-direction: column;
-  padding: 10px 0;
-  justify-content: space-between;
-  border-right: 1px solid rgba(200, 200, 200, 0.3);
-
-  .nav-list {
-    list-style-type: none;
-    padding-left: 15px;
-    margin-bottom: 0;
-    font-family: "Lato", sans-serif;
-
-    li {
-      padding: 9px 15px;
-      font-size: 15px;
-    }
-
-    .active {
-      background: green;
-    }
-  }
-  .top-div {
-    .image {
-      height: 100px;
-    }
-  }
-  .bottom-div {
-    padding: 15px 20px;
+  .nav-bar {
+    flex-grow: 1;
     display: flex;
     flex-direction: column;
+    padding: 10px 0;
+    justify-content: space-between;
+    border-right: 1px solid rgba(200, 200, 200, 0.3);
 
-    .social-links {
+    .nav-list {
       list-style-type: none;
-      padding: 0;
-      margin: 0;
-      display: flex;
-      justify-content: space-between;
-    }
+      padding-left: 15px;
+      margin-bottom: 0;
+      font-family: "Lato", sans-serif;
 
-    .copyright-info {
-      margin-top: 5px;
-      font-size: 7px;
-      align-self: center;
+      li {
+        padding: 9px 15px;
+        font-size: 15px;
+      }
+
+      .active {
+        background: green;
+      }
+    }
+    .top-div {
+      .image {
+        height: 100px;
+      }
+    }
+    .bottom-div {
+      padding: 15px 20px;
+      display: flex;
+      flex-direction: column;
+
+      .social-links {
+        list-style-type: none;
+        padding: 0;
+        margin: 0;
+        display: flex;
+        justify-content: space-between;
+      }
+
+      .copyright-info {
+        margin-top: 5px;
+        font-size: 7px;
+        align-self: center;
+      }
     }
   }
 }
